+++ Makefile.in
@@ -8,7 +8,7 @@ AT=@
 
 BUILD_DIRS = aprcl ulong_extras long_extras perm fmpz fmpz_vec fmpz_poly \
    fmpq_poly fmpz_mat fmpz_lll mpfr_vec mpfr_mat mpf_vec mpf_mat nmod_vec nmod_poly \
-   mpoly nmod_mpoly fmpz_mpoly fmpq_mpoly\
+   threadpool mpoly nmod_mpoly fmpz_mpoly fmpq_mpoly\
    nmod_poly_factor arith mpn_extras nmod_mat fmpq fmpq_vec fmpq_mat padic \
    fmpz_poly_q fmpz_poly_mat nmod_poly_mat fmpz_mod_poly \
    fmpz_mod_poly_factor fmpz_factor fmpz_poly_factor fft qsieve \
