+++ threadpool/init.c
@@ -50,6 +50,7 @@ threadpool_Unlock:
     return NULL;
 }
 
+
 void threadpool_init(threadpool_t T, slong l)
 {
     slong i;
@@ -85,124 +86,4 @@ void threadpool_init(threadpool_t T, slong l)
             pthread_cond_wait(&D[i].sleep2, &D[i].mutex);
         pthread_mutex_unlock(&D[i].mutex);
     }
-}
-
-slong threadpool_size(threadpool_t T)
-{
-    slong ret;
-    pthread_mutex_lock(&T->mutex);
-    ret = T->length;
-    pthread_mutex_unlock(&T->mutex);
-    return ret;
-}
-
-slong threadpool_request(threadpool_t T, threadpool_threadhandle * out, slong requested)
-{
-    slong i, ret = 0;
-    tpentry_struct * D;
-
-    if (requested <= 0)
-        return 0;
-
-    pthread_mutex_lock(&T->mutex);
-
-    D = T->tdata;
-    if (T->length > 0)
-    {
-        for (i = 0; i < T->length; i++)
-        {
-            if (D[i].available == 1)
-            {
-                D[i].available = 0;
-                out[ret] = i;
-                ret++;
-                if (ret >= requested)
-                    break;
-            }
-        }
-    }
-
-    pthread_mutex_unlock(&T->mutex);
-
-    return ret;
-}
-
-void threadpool_wake(threadpool_t T, threadpool_threadhandle i, void (*f)(void*), void * a)
-{
-    tpentry_struct * D;
-
-    pthread_mutex_lock(&T->mutex);
-
-    FLINT_ASSERT(i < T->length);
-
-    D = T->tdata;
-
-    FLINT_ASSERT(D[i].available == 0); /* should not be trying to wake an available thread */
-
-    pthread_mutex_lock(&D[i].mutex);
-    D[i].working = 1;
-    D[i].fxn = f;
-    D[i].fxnarg = a;
-    pthread_cond_signal(&D[i].sleep1);
-    pthread_mutex_unlock(&D[i].mutex);
-
-    pthread_mutex_unlock(&T->mutex);
-}
-
-void threadpool_wait(threadpool_t T, threadpool_threadhandle i)
-{
-    tpentry_struct * D;
-
-    D = T->tdata;
-
-    FLINT_ASSERT(D[i].available == 0); /* should not be trying to wait on a available thread */
-
-    pthread_mutex_lock(&D[i].mutex);
-    while (D[i].working != 0)
-        pthread_cond_wait(&D[i].sleep2, &D[i].mutex);        
-    pthread_mutex_unlock(&D[i].mutex);
-
-    pthread_mutex_lock(&T->mutex);
-    D[i].available = 0;
-    pthread_mutex_unlock(&T->mutex);
-}
-
-void threadpool_giveback(threadpool_t T, threadpool_threadhandle i)
-{
-    tpentry_struct * D;
-
-    D = T->tdata;
-
-    FLINT_ASSERT(D[i].available == 0); /* should not be trying to giveback an available thread */
-
-    pthread_mutex_lock(&D[i].mutex);
-    FLINT_ASSERT(D[i].working == 0);
-    pthread_mutex_unlock(&D[i].mutex);
-
-    pthread_mutex_lock(&T->mutex);
-    D[i].available = 1;    
-    pthread_mutex_unlock(&T->mutex);
-}
-
-void threadpool_clear(threadpool_t T)
-{
-    slong i;
-    tpentry_struct * D = T->tdata;
-    slong l = T->length;
-
-    for (i = 0; i < l; i++)
-    {
-        pthread_mutex_lock(&D[i].mutex);
-        D[i].exit = 1;
-        pthread_cond_signal(&D[i].sleep1);
-        pthread_mutex_unlock(&D[i].mutex);
-        pthread_join(D[i].pth, NULL);
-        pthread_cond_destroy(&D[i].sleep2);
-        pthread_cond_destroy(&D[i].sleep1);
-        pthread_mutex_destroy(&D[i].mutex);
-    }
-    flint_free(D);
-    pthread_mutex_destroy(&T->mutex);
-    T->length = -1;
-    T->tdata = NULL;
 }
